Add sanity checks for feh profile fragment

+++ feh/.profile.d/feh.sh
@@ -1,5 +1,7 @@
 if type feh >/dev/null 2>&1; then
 	for theme in $(awk '{print $1}' ~/.config/feh/themes); do
-		ln -s $(which feh) "$HOME/.local/bin/$theme"
+		if ! test -L "$HOME/.local/bin/$theme"; then
+			ln -s $(which feh) "$HOME/.local/bin/$theme"
+		fi
 	done
 fi
